Put in IHDR chunk logic

+++ utils/mathutils.py
@@ -7,8 +7,8 @@ def crc_table(val, poly, little_endian=True):
     if not little_endian:
         oper = lambda x, y: x << y
 
-    c = ctypes.c_uint32(val)
-    p = ctypes.c_uint32(poly)
+    c = val
+    p = poly
 
     for i in range(0,8):
         if c & 1:
@@ -24,7 +24,7 @@ def crc(buf, poly, initial=0x00000000, little_endian=True):
     if not little_endian:
         oper = lambda x,y: x << y
 
-    shift = ctypes.uint32(initial)
+    shift = initial
     for val in buf:
         shift = crc_table((shift ^ val) & 0xff, poly, little_endian) ^ (oper(shift, 8))
